Minimal css changes
Altered a bit of css so that the website can keep its aspect ratio when a client is resizing the window. Also created a new .card-image class for the cards that will replace the card slots on the table. Doesn't work perfectly, but it resizes the card images based on the window size.