Testing How Game Will Run
Trying to figure out how the game will be run, and how a user will interact with the java to make it run. Currently having some difficulty, because it is hard for the website to continually update the java code as it is now. Also can't use more than one request at a time, so the main poker page can't be changed by the settings page. Whenever I try to change the chip counts as well as start the game at the same time, an error occurs.